+++ .gitignore
@@ -1,3 +1,4 @@
+dist/
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
